renamed interpretccfg into compilerBackend
